# services/api-gateway/docs/api-gateway-openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "TrioBank API Gateway",
        "version": "1.0.0",
        "description": "TrioBank API Gateway - Unified access point for all microservices. Handles authentication, rate limiting, and routing to Ledger, Account, and Card services.",
        "contact": {
            "name": "TrioBank Platform",
            "url": "https://triobank.example.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Development"
        },
        {
            "url": "https://api.triobank.example.com",
            "description": "Production"
        }
    ],
...
    "paths": {
...
        "/v1/accounts": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Get customer accounts",
                "description": "Lists all accounts for a customer with optional status filter",
                "operationId": "getCustomerAccounts",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "ACTIVE",
                                    "CLOSED"
                                ]
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account list retrieved"
                    },
                    "400": {
                        "description": "Missing customerId parameter"
                    }
                }
            },
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Create account",
                "description": "Opens a new account for a customer",
                "operationId": "createAccount",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAccountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Account created successfully"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/v1/accounts/{id}": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Get account details",
                "description": "Retrieves details of a specific account",
                "operationId": "getAccount",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account details retrieved"
                    },
                    "404": {
                        "description": "Account not found"
                    }
                }
            }
        },
        "/v1/accounts/{id}/status": {
            "patch": {
                "tags": [
                    "Account"
                ],
                "summary": "Change account status",
                "description": "Changes account status (ACTIVE ↔ CLOSED)",
                "operationId": "changeAccountStatus",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    },
                    "400": {
                        "description": "Cannot reopen closed account"
                    }
                }
            }
        },
        "/v1/accounts/{id}/configuration": {
            "patch": {
                "tags": [
                    "Account"
                ],
                "summary": "Update account configuration",
                "description": "Updates user-specific account settings (limits, notifications, etc)",
                "operationId": "updateAccountConfiguration",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration updated"
                    }
                }
            }
        },
        "/v1/products": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Get all products",
                "description": "Lists all available account products",
                "operationId": "getAllProducts",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product list retrieved"
                    }
                }
            }
        },
        "/v1/products/{code}": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Get product details",
                "description": "Retrieves details of a specific product",
                "operationId": "getProduct",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details retrieved"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT token from /auth/login/confirm endpoint"
            }
        },
        "schemas": {
...
            "CreateAccountRequest": {
                "type": "object",
                "required": [
                    "customerId",
                    "productCode",
                    "currency"
                ],
                "properties": {
                    "customerId": {
                        "type": "string"
                    },
                    "productCode": {
                        "type": "string",
                        "example": "CHECKING_TRY"
                    },
                    "currency": {
                        "type": "string",
                        "example": "TRY"
                    },
                    "configurations": {
                        "type": "object",
                        "additionalProperties": true
                    }
                }
            },
...
        }
    }
}